Add support for translations
When multiple languages are available in one appdata or desktop file,
instead of merging everything togheter, add multiple name/summary/
description tags with different xml:lang attributes.
To do so, we also need to switch from ConfigParser to GLib.KeyFile,
which also implies switching to the introspected bindings for
GdkPixbuf and Rsvg.

Signed-off-by: Richard Hughes <richard@hughsie.com>