Don't mark the web-apps as 'Proprietary'
These applications do not have a clear license and it is not really clear how a
license tag makes sense when a lot of the time there are a mix, or unknown
licenses in use. Rather than guess, just don't set a value so the UI can do the
right thing.